requirements_pc.txt
# ── PC AI Server Requirements ──────────────────────────────────────────────────
# Install: pip install -r requirements_pc.txt

pyzmq>=25.0.0
opencv-python>=4.7.0
numpy>=1.24.0
Pillow>=9.0.0

# AI / Detection (heavy, requires torch)
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# OCR
easyocr>=1.7.0
arabic-reshaper>=3.0.0
python-bidi>=0.4.0
